refactor: set base path for better ablsolute imports usage
