+++ src/@modules/DependencyFormatter/components/DependencyList/DependencyList.tsx
