+++ src/@modules/DependencyFormatter/components/DependencyList/index.ts
