+++ src/@modules/DependencyFormatter/components/DependencyListItem/DependencyListItem.tsx
