+++ src/@modules/DependencyFormatter/components/DependencyListItem/index.ts
