+++ src/@modules/DependencyFormatter/slice.ts
