+++ src/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { App } from '@modules/App';
 import './index.css';
 import * as serviceWorker from './serviceWorker';
-import { App } from './App';
 
 ReactDOM.render(
   <React.StrictMode>
